+++ project_jj_lib_part2/Vector.h
@@ -10,16 +10,16 @@
 class Vector
 {
     uint64_t *vec;
-    uint64_t capacity, index;
+    uint32_t capacity, index;
 
 public:
     Vector();
     ~Vector();
 
-    uint64_t size();
+    uint32_t size();
     void push_back(uint64_t x);
     bool search(uint64_t x);
-    int getIndex(uint64_t x);
+    uint32_t getIndex(uint64_t x);
     void clear();
 
     uint64_t operator[] (uint64_t x);
